Added hashing for a key
Got rid of the line number.  It wasn’t proving to be especially useful.
 Replaced it with an MD5 hash value that should (hopefully) uniquely
identify the variant.  This can be described as the hex value for
MD5(chromosome:position/ref/alt1,alt2).  Note that a trailing slash
will be left if there are no alternate alleles listed (this is
irrelevant, as Seattleseq will leave the line off, but it is important
for a split VCF).  I will adjust the SQL scripts to key off this in the
very near future (this should hopefully make the joining of tables more
efficient)